requirements.txt
agno>=1.0.0
firecrawl>=0.1.0
gradio>=4.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0